backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
Pillow==11.0.0
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
python-decouple==3.8
django-apscheduler==0.6.2
openpyxl==3.1.2
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.10
dj-database-url==3.0.1
requests==2.31.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
reportlab==4.0.7
pdfminer.six==20231228